[RELEASE] 0.15.3 (#4981)
* ProgressBar for DataAssistant RuleBasedProfiler computations. (#4918)

* [MAINTENANCE] Ensure that code style scripts in CI/CD exit early on failure (#4921)

* chore: update pipeliens

* chore: remove scope check from pipeline

* [BUGFIX] check contrib requirements (#4922)

* Add check that requirements is a list, but don't crash if it's not

* Make requirements for icd_ten_category expectation a list

* [BUGFIX] Enables successful parsing of test cases for multi-table expectations (#4906)

* Remove pointless _generate_expectation_tests wrapper method and update docstring on generate_expectation_tests

* Remove accepting 'return_only_gallery_example's arg from run_diagnostics method

* Update build_gallery.py script to receive --no-core --no-contrib and arbitrary Expectation list

* Use phrase 'Has a valid library_metadata object'

* Update ExpectationTestDiagnostics to have include_in_gallery

* Update _get_metric_list to accept expectation_config instead of executed_test_cases

* Update ExpectationTestDiagnostics to include validation_result and error_diagnostics

* Delete _execute_test_examples, _choose_example, _instantiate_example_validation_results, and ExecutedExpectationTestCase

* Reformat with black

* Update run_diagnostics to determine maturity level based on checks passed

* Update evaluate_json_test_cfe to accept raise_exception and return a tuple

* Update _get_test_results to include more in ExpectationErrorDiagnostics via evaluate_json_test_cfe

* Add backend_test_result_counts to ExpectationDiagnostics and use in helpers

* Reformat with black

* Remove unused imports (flake8)

* Fix fix tests

* Update asserts at end of creating_custom_expectations/expect_xxx.py

* Add some print statements to generate_expectation_tests when get_test_validator_with_data has a problem

* test setup

* fixes diagnostics for multi-table expectations

* wrap tmp_dir -> abspath in func

* apply  to test_expectations/test_expectations_cfe

* docstring


* [MAINTENANCE] Remove BatchRequest from Rule-Based Profiler Configuration and from Builder Constructor Arguments (#4927)

* [BUGFIX] Add missing events to schema (#4917)

* [MAINTENANCE] Improve Altair plotting extensibility (#4923)

* Comments on altair documentation

* Predicate BinaryExpression type hint

* Make default theme and enum as well

* Pass custom config to altair

* Bugfix using nested_update

* Add tests that test notebook execution

* Add failing test

* Move opacity into theme, rename variable

* Vanquish tooltip and point_color_condition parameters

* [FEATURE] new checksum expectation (#4657)

* [FEATURE] code for new checksum expectation

* [FEATURE] code for new checksum expectation

* initial code for checksum expectation

* linting & library_metadata updates


* [BUGFIX] Update helper to add explicit alias to subqueries for SQLA version < 1.4 (#4660)

* Update helper to add explicit alias to subqueries for SQLA version < 1.4

Implicit conversion of a nested select into a subquery failed when
running on SQLA 1.3 against Postgres - update the existing helper to
also handle older supported versions of SQLA.

* Update util.py


* [BUGFIX] Fix clickhouse same-alias issue (#4389)

* Fix broken link for checklist (#4932)

* [MAINTENANCE] Remove DataContext from DataAssistant  (#4931)

* [MAINTENANCE] Add condition for custom checks in great_expectations pipelines

* Move general data splitting tasks to abstract base class (#4942)

* [MAINTENANCE] Add test to check for missing usage events (#4933)

* [FEATURE] Provide ability to combine lists of ExpectationConfiguration objects into flexible ExpectationSuite containers (#4943)

* [MAINTENANCE] Move splitter related taxi integration test fixtures (#4947)

* Update SLACK_GUIDELINES.md

Updating some language in Slack Guidelines

* added how to ask a question link

* cleanup (#4949)

* [MAINTENANCE] Rearrange modules for better reusability (#4955)

* [MAINTENANCE] Add timeout to great_expectations pipeline stages to prevent false positive build failures

* [FEATURE] Enable self-initializing capabilities for `ExpectColumnProportionOfUniqueValuesToBeBetween` (#4929)

* feat: init commit

* chore: misc changes per convo with Alex

* feat: finish initial impl

* feat: finish impl after convo with Alex

* chore: update after review

* clean up (#4959)

* [FEATURE] Enable support for plotting both Table and Column charts in `VolumeDataAssistant` (#4930)

* feat: init commit

* feat: continue chugging along

* feat: get both types of charts to work

* chore: only update relevant kwargs in df

* feat: add subtitle support

* feat: create predicate helper func

* chore: update type hint

* chore: bold subtitle

* chore: work on cleaning up vconcat

* feat: continue impl

* feat: get both prescriptive and descriptive working

* chore: delete unnecessary import

* refactor: further cleanup

* chore: shrink charts some more

* refactor: rename private method

* chore: add docstrings

* feat: add include/exclude column names lists

* fix: correct method calls

* fix: fix assertion around include/exclude columns

* chore: update styling of charts

* chore: misc changes per Nathan review

* [BUGFIX] Use `monkeypatch` to set a consistent bootstrap seed in tests (#4960)

* feat: start impl

* chore: finishing touches

* fix: remedy typo in test

* feat: update test

* chore: revert changes in utils

* chore: add comment

* pin cryptography package (#4963)

pin cryptography package (#4963)

* [FEATURE] BigQuery Temp Table Support (#4925)

* [FEATURE] BigQuery Temp Table Support (#4925)

* [FEATURE] Registry for DataAssistant classes with ability to execute from DataContext by registered name (#4966)

* [MAINTENANCE] Type annotate relevant functions with `-> None` (per PEP 484) (#4969)

* feat: run script to type annotate

* chore: update threshold

* Enable RuleBasedProfiler components to be serializable. (#4972)

* [BUGFIX] extras_require (#4968)

* Remove azure from requirements-dev-sqlalchemy.txt

* Update get_extras_require func to strip comments and include sqlalchemy for some keys

* [MAINTENANCE] Handle edge cases where `false_positive_rate` is not in range [0, 1] or very close to bounds (#4946)

* Warn and use NP_EPSILON if false_positive_rate <= 0, raise ValueError if false_positive_rate >= 1

* Add failing tests for corner cases

* Passing tests for 0 and 1 false_positive_rate

* Add tests for very small false_positive_rates

* Return type is already validated as float

* Use custom ProfilerExecutionError rather than ValueError

* Use 1-NP_EPSILON as an upper bound

* Pass variables to quentin fixture to set random seed

* Bugfix setting wrong parameter

* Set object attribute as well

* Unable to access the actual false_positive_rate used as it is private

* Use floats instead of ints

* Update type hints

* [MAINTENANCE] fix a typo  (#4974)

* [FEATURE] Enable self-intializing capabilities for `ExpectColumnValuesToMatchRegex`/`ExpectColumnValuesToNotMatchRegex` (#4958)

* feat: start impl

* test: start writing alice test

* feat: misc updates per discussion with Alex

* test: update test regexes

* feat: update other expectation

* chore: update fixtures

* chore: type hint

* [BUGFIX] Fix broken packaging test and update dgtest-overrides

* [FEATURE] Provide "estimation histogram" ParameterBuilder output details . (#4975)

* [FEATURE] Enable self-initializing ExpectColumnValuesToMatchStrftimeFormat

* release prep

+++ tests/test_utils.py
@@ -31,7 +31,6 @@ from great_expectations.data_context.util import (
     build_store_from_config,
     instantiate_class_from_config,
 )
-from great_expectations.datasource.data_connector import InferredAssetSqlDataConnector
 from great_expectations.execution_engine import SqlAlchemyExecutionEngine
 
 logger = logging.getLogger(__name__)
@@ -463,6 +462,24 @@ def get_snowflake_connection_url() -> str:
     return f"snowflake://{sfUser}:{sfPswd}@{sfAccount}/{sfDatabase}/{sfSchema}?warehouse={sfWarehouse}"
 
 
+def get_bigquery_table_prefix() -> str:
+    """Get table_prefix that will be used by the BigQuery client in loading a BigQuery table from DataFrame.
+
+    Reference link
+        https://cloud.google.com/bigquery/docs/samples/bigquery-load-table-dataframe
+
+    Returns:
+        String of table prefix, which is the gcp_project and dataset concatenated by a "."
+    """
+    gcp_project = os.environ.get("GE_TEST_GCP_PROJECT")
+    if not gcp_project:
+        raise ValueError(
+            "Environment Variable GE_TEST_GCP_PROJECT is required to run BigQuery integration tests"
+        )
+    bigquery_dataset = os.environ.get("GE_TEST_BIGQUERY_DATASET", "test_ci")
+    return f"""{gcp_project}.{bigquery_dataset}"""
+
+
 def get_bigquery_connection_url() -> str:
     """Get bigquery connection url from environment variables.
 
@@ -489,6 +506,45 @@ class LoadedTable:
     inserted_dataframe: pd.DataFrame
 
 
+def load_and_concatenate_csvs(
+    csv_paths: List[str],
+    load_full_dataset: bool = False,
+    convert_column_names_to_datetime: Optional[List[str]] = None,
+) -> pd.DataFrame:
+    """Utility method that is used in loading test data into a pandas dataframe.
+
+    It includes several parameters used to describe the output data.
+
+    Args:
+        csv_paths: list of paths of csvs to write, can be a single path. These should all be the same shape.
+        load_full_dataset: if False, load only the first 10 rows.
+        convert_column_names_to_datetime: List of column names to convert to datetime before writing to db.
+
+    Returns:
+        A pandas dataframe concatenating data loaded from all csvs.
+    """
+
+    if convert_column_names_to_datetime is None:
+        convert_column_names_to_datetime = []
+
+    import pandas as pd
+
+    dfs: List[pd.DataFrame] = []
+    for csv_path in csv_paths:
+        df = pd.read_csv(csv_path)
+        for column_name_to_convert in convert_column_names_to_datetime:
+            df[column_name_to_convert] = pd.to_datetime(df[column_name_to_convert])
+        if not load_full_dataset:
+            # Improving test performance by only loading the first 10 rows of our test data into the db
+            df = df.head(10)
+
+        dfs.append(df)
+
+    all_dfs_concatenated: pd.DataFrame = pd.concat(dfs)
+
+    return all_dfs_concatenated
+
+
 def load_data_into_test_database(
     table_name: str,
     connection_string: str,
@@ -513,41 +569,24 @@ def load_data_into_test_database(
             same prefix.
 
     Returns:
-        For convenience, the pandas dataframe that was used to load the data.
+        LoadedTable which for convenience, contains the pandas dataframe that was used to load the data.
     """
     if csv_path and csv_paths:
         csv_paths.append(csv_path)
     elif csv_path and not csv_paths:
         csv_paths = [csv_path]
 
-    if convert_colnames_to_datetime is None:
-        convert_colnames_to_datetime = []
+    all_dfs_concatenated: pd.DataFrame = load_and_concatenate_csvs(
+        csv_paths, load_full_dataset, convert_colnames_to_datetime
+    )
 
     if random_table_suffix:
         table_name: str = f"{table_name}_{str(uuid.uuid4())[:8]}"
 
-    import pandas as pd
-
-    print("Generating dataframe of all csv data")
-    dfs: List[pd.DataFrame] = []
-    for csv_path in csv_paths:
-        df = pd.read_csv(csv_path)
-        for colname_to_convert in convert_colnames_to_datetime:
-            df[colname_to_convert] = pd.to_datetime(df[colname_to_convert])
-        if not load_full_dataset:
-            # Improving test performance by only loading the first 10 rows of our test data into the db
-            df = df.head(10)
-
-        dfs.append(df)
-
-    all_dfs_concatenated: pd.DataFrame = pd.concat(dfs)
-
     return_value: LoadedTable = LoadedTable(
         table_name=table_name, inserted_dataframe=all_dfs_concatenated
     )
-
     connection = None
-
     if sa:
         engine = sa.create_engine(connection_string)
     else:
@@ -556,23 +595,57 @@ def load_data_into_test_database(
             "install optional sqlalchemy dependency for support."
         )
         return return_value
-    try:
-        connection = engine.connect()
-        print(f"Dropping table {table_name}")
-        connection.execute(f"DROP TABLE IF EXISTS {table_name}")
-        print(f"Creating table {table_name} and adding data from {csv_paths}")
-        all_dfs_concatenated.to_sql(
-            name=table_name, con=engine, index=False, if_exists="append"
+    if engine.dialect.name.lower() == "bigquery":
+        # bigquery is handled in a special way
+        load_data_into_test_bigquery_database_with_bigquery_client(
+            dataframe=all_dfs_concatenated, table_name=table_name
         )
         return return_value
-    except SQLAlchemyError as e:
-        logger.error(
-            """Docs integration tests encountered an error while loading test-data into test-database."""
+    else:
+        try:
+            connection = engine.connect()
+            print(f"Dropping table {table_name}")
+            connection.execute(f"DROP TABLE IF EXISTS {table_name}")
+            print(f"Creating table {table_name} and adding data from {csv_paths}")
+            all_dfs_concatenated.to_sql(
+                name=table_name, con=engine, index=False, if_exists="append"
+            )
+            return return_value
+        except SQLAlchemyError as e:
+            logger.error(
+                """Docs integration tests encountered an error while loading test-data into test-database."""
+            )
+            raise
+        finally:
+            connection.close()
+            engine.dispose()
+
+
+def load_data_into_test_bigquery_database_with_bigquery_client(
+    dataframe: pd.DataFrame, table_name: str
+) -> None:
+    """
+    Loads dataframe into bigquery table using BigQuery client. Follows pattern specified in the GCP documentation here:
+        - https://cloud.google.com/bigquery/docs/samples/bigquery-load-table-dataframe
+    Args:
+        dataframe (pd.DataFrame): DataFrame to load
+        table_name (str): table to load DataFrame to. Prefix containing project and dataset are loaded
+                        by helper function.
+    """
+    prefix: str = get_bigquery_table_prefix()
+    table_id: str = f"""{prefix}.{table_name}"""
+    from google.cloud import bigquery
+
+    gcp_project: Optional[str] = os.environ.get("GE_TEST_GCP_PROJECT")
+    if not gcp_project:
+        raise ValueError(
+            "Environment Variable GE_TEST_GCP_PROJECT is required to run BigQuery integration tests"
         )
-        raise
-    finally:
-        connection.close()
-        engine.dispose()
+    client: bigquery.Client = bigquery.Client(project=gcp_project)
+    job: bigquery.LoadJob = client.load_table_from_dataframe(
+        dataframe, table_id
+    )  # Make an API request.
+    job.result()  # Wait for the job to complete
 
 
 def clean_up_tables_with_prefix(connection_string: str, table_prefix: str) -> List[str]:
